[Internal] Prepare `ApiClient` to replace `DatabricksClient`. (#1185)
## What changes are proposed in this pull request?

`ApiClient` provides always all the information needed to replace
`DatabricksClient` in the service client. The only thing missing is
access to the Host and Account ID. This PR add these to the `ApiClient`
config. In preparation to the migration.

## How is this tested?

No test, the new function will be tested indirectly as soon as the
migration is completed.

NO_CHANGELOG=true