Stop recommending opening an issue in case of unknown error (#1189)
## What changes are proposed in this pull request?

This PR refactors the error parsing logic so that the SDK stop
recommending reporting an issue in case of unexpected error. The
rationale is that we do not have control over the error types that
middleware might return. It is thus not possible to reliably identify if
a parsing issue is due to a limitation of the SDK or to the error
itself.

This PR also slightly simplifies the code. 

## How is this tested?

Normal CI.