[Feature] Implement U2M Authentication in the Go SDK (#1108)
## What changes are proposed in this pull request?

This PR moves logic about U2M OAuth login from the CLI to the Go SDK.
This eliminates a cyclic dependency between the SDK and CLI for
interacting with the OAuth token cache and enables U2M support directly
in the Go SDK without need for the CLI to be installed.

Most of this code is carried over from the CLI, but I have made specific
refactors to generalize it where needed.

Currently, the token cache key follows a specific structure:
- `https://<accounts-host>/oidc/accounts/<account-id>` for account-based
sessions
- `https://<workspace host>` for workspace-based sessions
This can be generalized to allow callers to cache tokens in other
manners. For example, users may want to cache tokens per principal or
per set of OAuth scopes for their own OAuth applications.

Additionally, products should be able to use an isolated token cache or
a token cache backed by a different storage medium, like a database.
This is simple to do by introducing a token cache interface.
Implementers can define their own credential strategy and reuse the
PersistentAuth type to handle the negotiation.

## How is this tested?

Carried over tests from the CLI.

More tests are forthcoming, once a decision is made on this approach.