Tolerate trailing slash (#1211)
## What changes are proposed in this pull request?

A recent refactor of U2M logic to the SDK introduced a small bug where a
hostname with a trailing slash would be rejected when using `databricks
auth login`, e.g. `databricks auth login --host
https://myworkspace.cloud.databricks.com/`. This happens because
`AccountOAuthArgument` and `WorkspaceOAuthArgument` both validate that
the provided host does not have a trailing slash. Instead, we can
tolerate hostnames ending with a slash by using the
`CanonicalHostName()` method of `Config` to normalize the hostname,
removing the unneeded trailing slash before passing the normalized
hostname to U2M auth components. I have made a similar fix to
`getOidcEndpoints`.

## How is this tested?
Unit tests are added to verify that `getOauthArgument` on `Config` works
with normalized hostnames, and same for `getOidcEndpoints`.