Allow localhost on u2m auth (#1233)
## What changes are proposed in this pull request?

This PR relaxes the validation performed on the host URL to allow `http`
URLs if the URL points to localhost.

This is needed to add acceptance tests in the Databricks CLI for the
`databricks auth login` command. For context: The acceptance test
framework in the CLI has the ability to spawn a testserver on localhost
and record API requests sent to it. We can use the testserver to
validate that the correct API requests are being sent during `auth
login`, and that the command does not fail.

Prompted by
https://github.com/databricks/cli/pull/2988#discussion_r2154081183 where
I almost let a big regression pass.

NO_CHANGELOG=true

## How is this tested?

Unit test