Start decoupling `ApiClient` and `DatabricksClient`. (#1245)
## What changes are proposed in this pull request?

The SDK has two clients which serve very similar purposes: `ApiClient`
and `DatabricksClient`. This PR is a first step to pay to merge these
two client into a single consistent client. Specifically, it brings some
of the functionality from `DatabricksClient` in `ApiClient` in
preparation for a change of code generation logic.

This PR also slightly changes the way the `ApiClient` default parameter
values are set by moving the logic in its constructor.

## How is this tested?

Unit and integration tests.