Update `u2m` port selection mechanism to start from 8020. (#1269)
## What changes are proposed in this pull request?

This PR replaces the port selection mechanism in `U2M` to start from
`8020`, and fallback incrementally (`8021`, `8022`, and so on... until
8040) if a port is not free. The PR also provides an environment/config
variable to specify a specific port for the OAuth callback. If set, that
port disables the fallback mechanism.

This change fixes a bug in the Databricks CLI which is only allowlisted
for port `8020` at the moment.

## How is this tested?

Unit tests