Fix: change default HTTP timeout to 60s (#1301)
## What changes are proposed in this pull request?
This PR fixes the default configuration for HTTP timeout. A couple of
months ago we increased the timeout to 60s, but that was a no-op because
of this line that sets it to 30 seconds.

I caught it after noticing a couple of `request timed out after 30s of
inactivity` during testing.

## How is this tested?
Existing unit test.

NO_CHANGELOG=true