[Fix] Support for Spark versions with multiple Scala versions
This is a fix for https://github.com/databricks/terraform-provider-databricks/issues/5218.
The issue is cased by the fact that `Scala` selector had default value of `2.12` (for
Terraform only), and as result was filtering out 17.x versions that are released only with
Scala 2.13.

Fix is to relax the default value, but handle the case when there could be 2.12 and 2.13
versions for the same DBR version (16.x only).

Also bumped staticcheck to 0.6.1 to support compilation with Go 1.25

Signed-off-by: Alex Ott <alexey.ott@databricks.com>