Remove Billing API test (#1368)
## What changes are proposed in this pull request?
Remove Billing API test. 

This test recently started to time-out after backend changes which
caused the API to return slower.
Given that the test has very low benefit to the SDKs and it takes very
long time to run, we can remove it.
What this test is testing is handling of APIs which return a raw text
response, but this is also covered by Files tests.

## How is this tested?
N/A

NO_CHANGELOG=true